+++ CHANGELOG.md
@@ -7,3 +7,11 @@ and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## 0.1.0 - 2020-08-17
+
+### Added
+
+- Collection of identity and access
+- Collection of repositories
+- Integration setup documentation